Produce a data frame for one signal.

Obtains data for selected date ranges for all geographic regions of the United States. Available data sources and signals are documented in the COVIDcast signal documentation. Most (but not all) data sources are available at the county level, but the API can also return data aggregated to metropolitan statistical areas, hospital referral regions, or states, as desired, by using the geo_type argument. View vignette("covidcastR") for detailed example usage.

covidcast_signal(
  data_source,
  signal,
  start_day = NULL,
  end_day = NULL,
  geo_type = c("county", "hrr", "msa", "dma", "state"),
  geo_value = "*"
)

Arguments

data_source
String identifying the data source to query. See the signal documentation for a list of available sources and signals.

signal
String identifying the signal from that source to query.

start_day
Query data beginning on this date. Date object, or string in the form YYYYMMDD. If start_day is NULL, defaults to first day data is available for this signal.

end_day
Query data up to this date, inclusive. Date object or string in the form YYYYMMDD. If end_day is NULL, defaults to the most recent day data is available for this signal.

geo_type
The geography type for which to request this data, such as "county" or "state". Available types are described in the COVIDcast signal documentation. Defaults to "county".

geo_value
Which geography to return. The default, "*", fetches all geographies. To fetch a specific geography, specify its ID as a string.

Value

Data frame with matching data. Contains geo_value, time_value, direction, value, stderr, and sample_size columns. geo_value identifies the location, such as a state name or county FIPS code; time_value contains Date objects. value is the signal quantity requested and stderr its standard error if available. sample_size indicates the sample size available in that geography on that day; sample size may not be available. direction uses a local linear fit to estimate whether the signal in this region is currently increasing or decreasing. Consult the signal documentation for more details.

Details

Downloading large amounts of data may be slow, so this function prints messages for each day of data it downloads. To suppress these, use base::suppressMessages(), as in suppressMessages(covidcast_signal("fb-survey", ...)).

References

COVIDcast API documentation: https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html

COVIDcast public map: https://covidcast.cmu.edu/

Examples

if (FALSE) {
## fetches all counties from 2020-05-10 to the most recent available data:
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510")
## fetches all counties on just 2020-05-10 and no other days
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510",
                 end_day = "20200510")
## fetch all states on 2020-05-10, 2020-05-11, 2020-05-12
covidcast_signal("fb-survey", "raw_cli", start_day = "20200510",
                 end_day = "20200512", geo_type = "state")
## fetch all available data for just Pennsylvania
covidcast_signal("fb-survey", "raw_cli", geo_type = "state",
                 geo_value = "pa")
## fetch all available data in Pittsburgh metropolitan area (identified by
## CBSA ID)
covidcast_signal("fb-survey", "raw_cli", geo_type = "msa",
                 geo_value = "38300")
}